llm-service/requirements.txt
# LLM Service Requirements
# Python dependencies for vLLM wrapper service

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client
httpx>=0.26.0

# Configuration
pyyaml>=6.0.1

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# vLLM inference engine
vllm>=0.2.7

# For Korean language support
transformers>=4.36.0
